requirements.txt
streamlit
python-dotenv
PyMuPDF
python-docx
fpdf
google-generativeai
